requirements.txt
# setup.sh
streamlit
mysql-connector-python
